+++ include/stage/effects/bridgeDebris.h
@@ -18,6 +18,6 @@ typedef struct EffectBridgeDebris_
 // FUNCTIONS
 // --------------------
 
-EffectBridgeDebris *EffectBridgeDebris__Create(fx32 x, fx32 y, fx32 velX, fx32 velY, u32 type);
+EffectBridgeDebris *EffectBridgeDebris__Create(fx32 x, fx32 y, fx32 velX, fx32 velY, u8 type);
 
 #endif // RUSH2_EFFECT_BRIDGE_DEBRIS_H
