+++ include/stage/effects/coralDebris.h
@@ -18,6 +18,6 @@ typedef struct EffectCoralDebris_
 // FUNCTIONS
 // --------------------
 
-EffectCoralDebris *EffectCoralDebris__Create(fx32 x, fx32 y, fx32 velX, fx32 velY, u32 type);
+EffectCoralDebris *EffectCoralDebris__Create(fx32 x, fx32 y, fx32 velX, fx32 velY, u8 type);
 
 #endif // RUSH2_EFFECT_CORAL_DEBRIS_H
