+++ include/stage/effects/steam.h
@@ -18,7 +18,7 @@ typedef struct EffectSteam_
 // FUNCTIONS
 // --------------------
 
-EffectSteam *EffectSteamDust__Create(u32 type, fx32 x, fx32 y, fx32 velX, fx32 velY);
-EffectSteam *EffectSteamEffect__Create(u32 type, fx32 x, fx32 y, fx32 velX, fx32 velY, s32 timer);
+EffectSteam *EffectSteamDust__Create(u8 type, fx32 x, fx32 y, fx32 velX, fx32 velY);
+EffectSteam *EffectSteamEffect__Create(u8 type, fx32 x, fx32 y, fx32 velX, fx32 velY, s32 timer);
 
 #endif // RUSH2_EFFECT_STEAM_H
